Adopt simplified configuration
