+++ website/package.json
@@ -23,8 +23,7 @@
   "devDependencies": {
     "gatsby": "^2.15.0",
     "gatsby-plugin-no-sourcemaps": "^2.0.2",
-    "gatsby-theme-ocular": "^1.0.0",
-    "gh-pages": "^2.1.0",
-    "sharp": "^0.23.0"
+    "gatsby-theme-ocular": "^1.1.0-alpha.2",
+    "gh-pages": "^2.1.0"
   }
 }
